fixed bug due to automatic box and unboxing  ....
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@609623 13f79535-47bb-0310-9956-ffa450edef68

+++ upnp/samples/tv/src/main/java/org/apache/felix/upnp/sample/tv/TvDevice.java
@@ -253,7 +253,7 @@ public class TvDevice implements UPnPDevice,UPnPEventListener,ServiceListener  {
 			LinkedDevices.add(deviceId);
 		if (deviceId.indexOf("Clock") != -1){
 				Long time = (Long) events.get("Time");
-				clockTime = new Date(time).toString();				
+				clockTime = new Date(time.longValue()).toString();				
 		}
 		else if (deviceId.indexOf("AirCon") != -1)
 				airconTemp = (String) events.get("Temp");
